Switch publication of blog to this repo.
This (should) switch the publication of the blog to the gh-pages branch
of _this_ repo.

It also remove the ssh key that allows to upload to my fork and replace
it for an ssh key that can push to this repo.

The key can be revoked at anytime and should is used only for this
operation.